# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCPJam API",
    "version": "1.0.0-preview",
    "description": "Programmatic access to MCP servers saved in your MCPJam projects — live diagnostics (validate, inspect, export) and operations: call tools, render prompts, run eval suites asynchronously and poll their results, and import OAuth tokens.\n\n**The API is in preview**: the surface may change while we finish the design. Error `code` values are stable; error `message` strings are not. Write clients that ignore unknown response fields.",
    "contact": {
      "name": "MCPJam",
      "url": "https://github.com/MCPJam/inspector/issues"
    }
  },
  "servers": [
    {
      "url": "https://app.mcpjam.com/api/v1",
      "description": "Hosted MCPJam"
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/projects/{projectId}/eval-runs": {
      "post": {
        "operationId": "createEvalRun",
        "tags": ["Eval runs"],
        "summary": "Create an eval run (async)",
        "description": "Creates a suite run from an existing `suiteId` (rerun) and/or inline `tests`, then **detaches execution and responds `202` immediately** with the `runId`. Validation and quota errors surface on this request; poll `GET /eval-runs/{runId}` for progress. The run appears live in the hosted UI Runs tab, tagged `source: \"api\"`.\n\nA bare `suiteId` with no inline tests reruns the suite as configured. Per-organization concurrency is capped (default 2 concurrent runs); exceeding it returns `429` with `details.reason: \"CONCURRENT_RUN_LIMIT\"`.",
        "parameters": [{ "$ref": "#/components/parameters/projectId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EvalRunCreateRequest" },
              "example": {
                "suiteName": "smoke",
                "serverIds": ["srv_abc123"],
                "tests": [
                  {
                    "title": "echo works",
                    "query": "Use the echo tool to say hi",
                    "runs": 1,
                    "model": "anthropic/claude-haiku-4.5",
                    "provider": "anthropic",
                    "expectedToolCalls": [
                      { "toolName": "echo", "arguments": { "text": "hi" } }
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Run created; execution continues in the background.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EvalRunCreated" },
                "example": {
                  "runId": "run_abc123",
                  "suiteId": "suite_def456",
                  "status": "running",
                  "caseUpsert": {
                    "committed": [{ "name": "echo works" }],
                    "failed": []
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/ServerUnreachable" },
          "504": { "$ref": "#/components/responses/Timeout" }
        }
      }
    },
    "/projects/{projectId}/eval-runs/{runId}": {
      "get": {
        "operationId": "getEvalRun",
        "tags": ["Eval runs"],
        "summary": "Get run status",
        "description": "Run status, result, and summary. Poll until `status` is terminal (`completed`, `failed`, or `cancelled`).",
        "parameters": [
          { "$ref": "#/components/parameters/projectId" },
          { "$ref": "#/components/parameters/runId" }
        ],
        "responses": {
          "200": {
            "description": "The run.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EvalRun" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{projectId}/eval-runs/{runId}/iterations": {
      "get": {
        "operationId": "listEvalRunIterations",
        "tags": ["Eval runs"],
        "summary": "List run iterations",
        "description": "Per-iteration results: actual tool calls, structured token usage, and latency. Cursor-paginated.",
        "parameters": [
          { "$ref": "#/components/parameters/projectId" },
          { "$ref": "#/components/parameters/runId" },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size, 1–200. Defaults to 50.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque cursor from a previous response's `nextCursor`.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of iterations.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EvalIterationPage" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{projectId}/eval-runs/{runId}/iterations/{iterationId}/trace": {
      "get": {
        "operationId": "getEvalIterationTrace",
        "tags": ["Eval runs"],
        "summary": "Get an iteration trace",
        "description": "Full trace envelope for one iteration: conversation messages, expected-vs-actual tool call analysis, and spans. The shape is rich and may evolve — treat it as an open document. Returns `404` with `details.reason: \"TRACE_NOT_AVAILABLE\"` when the iteration finished without a stored trace.",
        "parameters": [
          { "$ref": "#/components/parameters/projectId" },
          { "$ref": "#/components/parameters/runId" },
          { "$ref": "#/components/parameters/iterationId" }
        ],
        "responses": {
          "200": {
            "description": "The trace envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Trace document: `messages`, `prompts` (per-prompt expected/actual tool-call analysis), `spans`, and more.",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{projectId}/eval-suites/{suiteId}/runs": {
      "get": {
        "operationId": "listEvalSuiteRuns",
        "tags": ["Eval runs"],
        "summary": "List a suite's runs",
        "description": "Recent runs for a suite, newest first.",
        "parameters": [
          { "$ref": "#/components/parameters/projectId" },
          { "$ref": "#/components/parameters/suiteId" },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum runs to return, 1–100. Defaults to 25.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent runs, newest first.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EvalRunPage" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "MCPJam API key (`sk_…`). Create one at [Settings → API keys](https://app.mcpjam.com/settings/api-keys). Guest sessions cannot use the API, and API keys cannot manage other API keys."
      }
    },
    "parameters": {
      "projectId": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "description": "ID of the hosted project that contains the server.",
        "schema": { "type": "string" }
      },
...
      "runId": {
        "name": "runId",
        "in": "path",
        "required": true,
        "description": "Eval run ID, as returned by `POST /eval-runs`.",
        "schema": { "type": "string" }
      },
...
    },
...
    "schemas": {
...
      "EvalRunCreateRequest": {
        "type": "object",
        "description": "Two valid shapes: `suiteId` (rerun an existing suite, optionally upserting inline `tests` into it), or `suiteName` + `tests` + `serverIds` (create a new suite and run it). Inline `tests` alone — without a `suiteId` or a `suiteName` — are rejected with `VALIDATION_ERROR`.",
        "anyOf": [
          { "required": ["suiteId"] },
          {
            "required": ["suiteName", "tests", "serverIds"],
            "properties": { "tests": { "minItems": 1 } }
          }
        ],
        "properties": {
          "suiteId": {
            "type": "string",
            "description": "Existing suite to rerun. A bare `suiteId` with no `tests` reruns the suite exactly as configured."
          },
          "suiteName": {
            "type": "string",
            "description": "Name for a new suite. Required (non-empty) when no `suiteId` is given."
          },
          "suiteDescription": { "type": "string" },
          "tests": {
            "type": "array",
            "maxItems": 100,
            "description": "Inline test cases to upsert into the suite before running.",
            "items": { "$ref": "#/components/schemas/EvalTestCase" }
          },
          "serverIds": {
            "type": "array",
            "minItems": 1,
            "description": "Servers (by ID) the run connects to. Required when creating a new suite; optional on reruns — when omitted, the run connects the suite's saved server selection (the set its snapshot references). A rerun of a suite with no saved selection is rejected with `VALIDATION_ERROR` (`details.reason: \"NO_SAVED_SERVER_SELECTION\"`).",
            "items": { "type": "string" }
          },
          "modelApiKeys": {
            "type": "object",
            "description": "Optional per-provider model API keys (e.g. `{ \"anthropic\": \"sk-ant-…\" }`). Falls back to your organization's configured providers when omitted.",
            "additionalProperties": { "type": "string" }
          },
          "notes": { "type": "string" },
          "passCriteria": {
            "type": "object",
            "properties": {
              "minimumPassRate": { "type": "number" }
            }
          },
          "iterationOverride": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "description": "Override the per-case `runs` count for this run only."
          }
        },
        "additionalProperties": true
      },
...
    },
...
  }
}